+++ app/examples/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="resourcecenterdemo">
+    package="animal.forest.chat">
 
     <uses-permission android:name="android.permission.INTERNET" />
