+++ app/examples/animal/forest/chat/util/PNFragmentImpl.java
@@ -1,0 +1,10 @@
+package animal.forest.chat.util;
+
+import com.pubnub.api.callbacks.SubscribeCallback;
+
+// tag::FRG-3[]
+public interface PNFragmentImpl {
+
+    SubscribeCallback provideListener();
+}
+// end::FRG-3[]
