+++ app/examples/animal/forest/chat/util/PNGlideModule.java
@@ -1,0 +1,9 @@
+package animal.forest.chat.util;
+
+import com.bumptech.glide.annotation.GlideModule;
+import com.bumptech.glide.module.AppGlideModule;
+
+@GlideModule
+public class PNGlideModule extends AppGlideModule {
+
+}
