--- app/examples/resourcecenterdemo/SplashActivity.java
@@ -1,24 +1,0 @@
-package resourcecenterdemo;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-public class SplashActivity extends AppCompatActivity implements Runnable {
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash);
-        new Handler().postDelayed(this, 100);
-    }
-
-    @Override
-    public void run() {
-        startActivity(new Intent(this, MainActivity.class));
-        finish();
-    }
-}
